requirements.txt
beautifulsoup4==4.8.0
numpy==1.19.5
pandas==0.25.1
python-rake==1.4.4
scikit_learn==1.0.1
sentence_transformers==2.1.0
streamlit==1.1.0
lxml==4.4.1
